# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ghana School Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/attendance:
    get:
      operationId: getAttendanceReport
      tags: [reports]
      summary: Attendance report by class and term
      parameters:
        - name: classId
          in: query
          required: false
          schema:
            type: integer
        - name: termId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Attendance report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceSummary"
  /reports/fees:
    get:
      operationId: getFeesReport
      tags: [reports]
      summary: Fee collection report by class and term
      parameters:
        - name: classId
          in: query
          required: false
          schema:
            type: integer
        - name: termId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Fee report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeesReport"
  /reports/results:
    get:
      operationId: getResultsReport
      tags: [reports]
      summary: Results/performance report by class and term
      parameters:
        - name: classId
          in: query
          required: false
          schema:
            type: integer
        - name: termId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Results report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClassResultSummary"
...
components:
  schemas:
...
    AttendanceSummary:
      type: object
      properties:
        studentId:
          type: integer
        studentName:
          type: string
        classId:
          type: integer
        className:
          type: ["string", "null"]
        totalDays:
          type: integer
        presentDays:
          type: integer
        absentDays:
          type: integer
        lateDays:
          type: integer
        excusedDays:
          type: integer
        attendanceRate:
          type: number
      required: [studentId, studentName, classId, totalDays, presentDays, absentDays, lateDays, excusedDays, attendanceRate]
...
    FeesReport:
      type: object
      properties:
        termName:
          type: ["string", "null"]
        totalExpected:
          type: number
        totalCollected:
          type: number
        totalPending:
          type: number
        collectionRate:
          type: number
        byClass:
          type: array
          items:
            type: object
            properties:
              className:
                type: string
              totalExpected:
                type: number
              totalCollected:
                type: number
              balance:
                type: number
              studentCount:
                type: integer
            required: [className, totalExpected, totalCollected, balance, studentCount]
        byFeeType:
          type: array
          items:
            type: object
            properties:
              feeTypeName:
                type: string
              totalExpected:
                type: number
              totalCollected:
                type: number
              balance:
                type: number
            required: [feeTypeName, totalExpected, totalCollected, balance]
      required: [totalExpected, totalCollected, totalPending, collectionRate, byClass, byFeeType]
    ClassResultSummary:
      type: object
      properties:
        className:
          type: string
        subjectName:
          type: ["string", "null"]
        averageScore:
          type: ["number", "null"]
        highestScore:
          type: ["number", "null"]
        lowestScore:
          type: ["number", "null"]
        passRate:
          type: ["number", "null"]
        studentCount:
          type: integer
      required: [className, studentCount]